Project: Simple Search Engine
https://hyperskill.org/projects/89

About
In this project you will create your own search engine, though it probably won’t compete with Google. Yours will be a simple search engine that processes a limited collection of data and searches it for a word or a phrase. It’ll be simple but cool.

Learning outcomes
Gain confidence in your ability to work with files and the console, and learn how to optimize the search process.